+++ src/Symfony/Component/HttpFoundation/CHANGELOG.md
@@ -1,6 +1,11 @@
 CHANGELOG
 =========
 
+3.1.0
+-----
+
+ * Added support for creating `JsonResponse` with a string of JSON data
+
 3.0.0
 -----